+++ forui/lib/src/widgets/button/button.dart
@@ -1,3 +1,4 @@
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
@@ -13,16 +14,16 @@ part 'button.style.dart';
 /// [FButton]s typically contain icons and/or a label. If the [onPress] and [onLongPress] callbacks are null, then this
 /// button will be disabled, it will not react to touch.
 ///
-/// The constants in [FButtonStyle] provide a convenient way to style a badge.
+/// The constants in [FButtonStyle] provide a convenient way to style a button.
 ///
 /// See:
 /// * https://forui.dev/docs/form/button for working examples.
-/// * [FButtonCustomStyle] for customizing a button's appearance.
+/// * [FButtonStyle] for customizing a button's appearance.
 class FButton extends StatelessWidget {
   /// The style. Defaults to [FButtonStyle.primary].
   ///
-  /// Although typically one of the pre-defined styles in [FButtonStyle], it can also be a [FButtonCustomStyle].
-  final FButtonStyle style;
+  /// Although typically one of the pre-defined styles in [FBaseButtonStyle], it can also be a [FButtonStyle].
+  final FBaseButtonStyle style;
 
   /// {@macro forui.foundation.FTappable.onPress}
   final VoidCallback? onPress;
@@ -96,15 +97,13 @@ class FButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final style = switch (this.style) {
-      final FButtonCustomStyle style => style,
+      final FButtonStyle style => style,
       Variant.primary => context.theme.buttonStyles.primary,
       Variant.secondary => context.theme.buttonStyles.secondary,
       Variant.destructive => context.theme.buttonStyles.destructive,
       Variant.outline => context.theme.buttonStyles.outline,
       Variant.ghost => context.theme.buttonStyles.ghost,
     };
-
-    final enabled = onPress != null || onLongPress != null;
 
     return FTappable.animated(
       focusedOutlineStyle: style.focusedOutlineStyle,
@@ -114,15 +113,10 @@ class FButton extends StatelessWidget {
       onPress: onPress,
       onLongPress: onLongPress,
       builder:
-          (_, data, child) => DecoratedBox(
-            decoration: switch ((enabled, data.hovered)) {
-              (true, false) => style.enabledBoxDecoration,
-              (true, true) => style.enabledHoverBoxDecoration,
-              (false, _) => style.disabledBoxDecoration,
-            },
-            child: child,
+          (_, states, _) => DecoratedBox(
+            decoration: style.boxDecoration.resolve(states),
+            child: FButtonData(style: style, states: states, child: child),
           ),
-      child: FButtonData(style: style, enabled: enabled, child: child),
     );
   }
 
@@ -141,51 +135,45 @@ class FButton extends StatelessWidget {
 
 /// A [FButton]'s style.
 ///
-/// A style can be either one of the pre-defined styles in [FButtonStyle] or a [FButtonCustomStyle]. The pre-defined
-/// styles are a convenient shorthand for the various [FButtonCustomStyle]s in the current context's [FButtonStyles].
-sealed class FButtonStyle {
+/// See [FButtonStyle] for more information.
+sealed class FBaseButtonStyle {}
+
+@internal
+enum Variant implements FBaseButtonStyle { primary, secondary, destructive, outline, ghost }
+
+/// A [FButton] style.
+///
+/// A style can be either one of the pre-defined styles or a [FButtonStyle]. The pre-defined styles are a convenient
+/// shorthand for the various [FButtonStyle]s in the current context's [FButtonStyles].
+class FButtonStyle extends FBaseButtonStyle with Diagnosticable, _$FButtonStyleFunctions {
   /// The button's primary style.
   ///
   /// Shorthand for the current context's [FButtonStyles.primary] style.
-  static const FButtonStyle primary = Variant.primary;
+  static const FBaseButtonStyle primary = Variant.primary;
 
   /// The button's secondary style.
   ///
   /// Shorthand for the current context's [FButtonStyles.secondary] style.
-  static const FButtonStyle secondary = Variant.secondary;
+  static const FBaseButtonStyle secondary = Variant.secondary;
 
   /// The button's destructive style.
   ///
   /// Shorthand for the current context's [FButtonStyles.destructive] style.
-  static const FButtonStyle destructive = Variant.destructive;
+  static const FBaseButtonStyle destructive = Variant.destructive;
 
   /// The button's outline style.
   ///
   /// Shorthand for the current context's [FButtonStyles.outline] style.
-  static const FButtonStyle outline = Variant.outline;
+  static const FBaseButtonStyle outline = Variant.outline;
 
   /// The button's ghost style.
   ///
   /// Shorthand for the current context's [FButtonStyles.ghost] style.
-  static const FButtonStyle ghost = Variant.ghost;
-}
+  static const FBaseButtonStyle ghost = Variant.ghost;
 
-@internal
-enum Variant implements FButtonStyle { primary, secondary, destructive, outline, ghost }
-
-/// A custom [FButton] style.
-class FButtonCustomStyle extends FButtonStyle with Diagnosticable, _$FButtonCustomStyleFunctions {
-  /// The box decoration for an enabled button.
+  /// The box decoration.
   @override
-  final BoxDecoration enabledBoxDecoration;
-
-  /// The box decoration for an enabled button when it is hovered over.
-  @override
-  final BoxDecoration enabledHoverBoxDecoration;
-
-  /// The box decoration for a disabled button.
-  @override
-  final BoxDecoration disabledBoxDecoration;
+  final FWidgetStateMap<BoxDecoration> boxDecoration;
 
   /// The focused outline style.
   @override
@@ -203,41 +191,31 @@ class FButtonCustomStyle extends FButtonStyle with Diagnosticable, _$FButtonCust
   @override
   final FButtonSpinnerStyle spinnerStyle;
 
-  /// Creates a [FButtonCustomStyle].
-  FButtonCustomStyle({
-    required this.enabledBoxDecoration,
-    required this.enabledHoverBoxDecoration,
-    required this.disabledBoxDecoration,
+  /// Creates a [FButtonStyle].
+  FButtonStyle({
+    required this.boxDecoration,
     required this.focusedOutlineStyle,
     required this.contentStyle,
     required this.iconContentStyle,
     required this.spinnerStyle,
   });
 
-  /// Creates a [FButtonCustomStyle] that inherits its properties from the given arguments.
-  FButtonCustomStyle.inherit({
+  /// Creates a [FButtonStyle] that inherits its properties from the given arguments.
+  FButtonStyle.inherit({
     required FTypography typography,
     required FStyle style,
-    required Color enabledBoxColor,
-    required Color enabledHoveredBoxColor,
-    required Color disabledBoxColor,
-    required Color enabledContentColor,
-    required Color disabledContentColor,
+    required FWidgetStateMap<Color> boxColor,
+    required FWidgetStateMap<Color> contentColor,
   }) : this(
-         enabledBoxDecoration: BoxDecoration(borderRadius: style.borderRadius, color: enabledBoxColor),
-         enabledHoverBoxDecoration: BoxDecoration(borderRadius: style.borderRadius, color: enabledHoveredBoxColor),
-         disabledBoxDecoration: BoxDecoration(borderRadius: style.borderRadius, color: disabledBoxColor),
+         boxDecoration: FWidgetStateMap(
+           boxColor.map(
+             (state, color) => MapEntry(state, BoxDecoration(borderRadius: style.borderRadius, color: color)),
+           ),
+         ),
          focusedOutlineStyle: style.focusedOutlineStyle,
-         contentStyle: FButtonContentStyle.inherit(
-           typography: typography,
-           enabled: enabledContentColor,
-           disabled: disabledContentColor,
-         ),
-         iconContentStyle: FButtonIconContentStyle(
-           enabledColor: enabledContentColor,
-           disabledColor: disabledContentColor,
-         ),
-         spinnerStyle: FButtonSpinnerStyle.inherit(enabled: enabledContentColor, disabled: disabledContentColor),
+         contentStyle: FButtonContentStyle.inherit(typography: typography, color: contentColor),
+         iconContentStyle: FButtonIconContentStyle(color: contentColor),
+         spinnerStyle: FButtonSpinnerStyle(color: contentColor),
        );
 }
 
@@ -255,22 +233,24 @@ class FButtonData extends InheritedWidget {
   }
 
   /// The button's style.
-  final FButtonCustomStyle style;
+  final FButtonStyle style;
 
-  /// True if the button is enabled.
-  final bool enabled;
+  /// The states.
+  ///
+  /// {@macro forui.foundation.tappable.builder}
+  final Set<WidgetState> states;
 
   /// Creates a [FButtonData].
-  const FButtonData({required this.style, required super.child, this.enabled = true, super.key});
+  const FButtonData({required this.style, required this.states, required super.child, super.key});
 
   @override
-  bool updateShouldNotify(covariant FButtonData old) => style != old.style || enabled != old.enabled;
+  bool updateShouldNotify(covariant FButtonData old) => style != old.style || !setEquals(states, old.states);
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
     properties
       ..add(DiagnosticsProperty('style', style))
-      ..add(FlagProperty('enabled', value: enabled, ifTrue: 'enabled'));
+      ..add(IterableProperty('states', states));
   }
 }