+++ forui/lib/src/widgets/button/button_content.dart
@@ -1,3 +1,4 @@
+import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
@@ -17,16 +18,13 @@ class Content extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final FButtonData(style: FButtonCustomStyle(:contentStyle), :enabled) = FButtonData.of(context);
+    final FButtonData(style: FButtonStyle(:contentStyle), :states) = FButtonData.of(context);
     return Padding(
       padding: contentStyle.padding,
       child: DefaultTextStyle.merge(
-        style: enabled ? contentStyle.enabledTextStyle : contentStyle.disabledTextStyle,
+        style: contentStyle.textStyle.resolve(states),
         child: FIconStyleData(
-          style: FIconStyle(
-            color: enabled ? contentStyle.enabledIconColor : contentStyle.disabledIconColor,
-            size: contentStyle.iconSize,
-          ),
+          style: FIconStyle(color: contentStyle.iconColor.resolve(states), size: contentStyle.iconSize),
           child: Row(
             mainAxisAlignment: MainAxisAlignment.center,
             spacing: 10,
@@ -46,15 +44,12 @@ class IconContent extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final FButtonData(:style, :enabled) = FButtonData.of(context);
+    final FButtonData(:style, :states) = FButtonData.of(context);
 
     return Padding(
       padding: style.iconContentStyle.padding,
       child: FIconStyleData(
-        style: FIconStyle(
-          color: enabled ? style.iconContentStyle.enabledColor : style.iconContentStyle.disabledColor,
-          size: style.iconContentStyle.size,
-        ),
+        style: FIconStyle(color: style.iconContentStyle.color.resolve(states), size: style.iconContentStyle.size),
         child: child,
       ),
     );
@@ -63,25 +58,21 @@ class IconContent extends StatelessWidget {
 
 /// [FButton] content's style.
 final class FButtonContentStyle with Diagnosticable, _$FButtonContentStyleFunctions {
-  /// The [TextStyle] when this button is enabled.
+  /// The [TextStyle].
+  ///
+  /// {macro forui.foundation.tappable.builder}
   @override
-  final TextStyle enabledTextStyle;
-
-  /// The [TextStyle] when this button is disabled.
-  @override
-  final TextStyle disabledTextStyle;
+  final FWidgetStateMap<TextStyle> textStyle;
 
   /// The padding. Defaults to `EdgeInsets.symmetric(horizontal: 16, vertical: 12.5)`.
   @override
   final EdgeInsetsGeometry padding;
 
-  /// The icon's color when this button is enabled.
+  /// The icon's color.
+  ///
+  /// {macro forui.foundation.tappable.builder}
   @override
-  final Color enabledIconColor;
-
-  /// The icon's color when this button is disabled.
-  @override
-  final Color disabledIconColor;
+  final FWidgetStateMap<Color> iconColor;
 
   /// The icon's size. Defaults to 20.
   @override
@@ -89,21 +80,22 @@ final class FButtonContentStyle with Diagnosticable, _$FButtonContentStyleFuncti
 
   /// Creates a [FButtonContentStyle].
   FButtonContentStyle({
-    required this.enabledTextStyle,
-    required this.disabledTextStyle,
-    required this.enabledIconColor,
-    required this.disabledIconColor,
+    required this.textStyle,
+    required this.iconColor,
     this.padding = const EdgeInsets.symmetric(horizontal: 16, vertical: 12.5),
     this.iconSize = 20,
   });
 
-  /// Creates a [FButtonContentStyle] that inherits its properties from the given [enabled] and [disabled].
-  FButtonContentStyle.inherit({required FTypography typography, required Color enabled, required Color disabled})
+  /// Creates a [FButtonContentStyle] that inherits its properties from the given [typography] and [color].
+  FButtonContentStyle.inherit({required FTypography typography, required FWidgetStateMap<Color> color})
     : this(
-        enabledTextStyle: typography.base.copyWith(color: enabled, fontWeight: FontWeight.w500, height: 1),
-        disabledTextStyle: typography.base.copyWith(color: disabled, fontWeight: FontWeight.w500, height: 1),
-        enabledIconColor: enabled,
-        disabledIconColor: disabled,
+        textStyle: FWidgetStateMap(
+          color.map(
+            (state, color) =>
+                MapEntry(state, typography.base.copyWith(color: color, fontWeight: FontWeight.w500, height: 1)),
+          ),
+        ),
+        iconColor: color,
       );
 }
 
@@ -113,23 +105,16 @@ final class FButtonIconContentStyle with Diagnosticable, _$FButtonIconContentSty
   @override
   final EdgeInsetsGeometry padding;
 
-  /// The icon's color when this button is enabled.
+  /// The icon's color.
+  ///
+  /// {macro forui.foundation.tappable.builder}
   @override
-  final Color enabledColor;
-
-  /// The icon's color when this button is disabled.
-  @override
-  final Color disabledColor;
+  final FWidgetStateMap<Color> color;
 
   /// The icon's size. Defaults to 20.
   @override
   final double size;
 
   /// Creates a [FButtonIconContentStyle].
-  const FButtonIconContentStyle({
-    required this.enabledColor,
-    required this.disabledColor,
-    this.padding = const EdgeInsets.all(7.5),
-    this.size = 20,
-  });
+  const FButtonIconContentStyle({required this.color, this.padding = const EdgeInsets.all(7.5), this.size = 20});
 }